+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mocha/Chai testing
 #### Changed
 - Package.json to support Mocha tests
-### v0.1.2 - 2019 07 01
+### v0.2.0 - 2019 07 01
 #### Added
 - Folders for all ESI groups with public endpoints
 #### Changed
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed constants declared as `let`s to `const`s
 - Added `getSettings` and `setSettings`
 - Datasource is now modifiable
-- Added Faction Warfare to Sovereignty
+- Added ESI groups from Faction Warfare to Sovereignty
 
 
 <!-- LINKS -->